requirements.txt
torchvision~=0.9.0
opencv~=4.0.1
pytest-shutil~=1.7.0
pytorch~=1.8.0
catboost~=0.25.1
numpy~=1.20.1
pillow~=8.2.0